move path provider into utils
